rearranged tests into seperate tests, and wrote most of the test for applicant api
