+++ applicant.html
@@ -31,5 +31,7 @@
             <dd id="like-food"></dd>
         </dl>
     </main>
+
+    <script type="module" src="./src/applicant.js"></script>
 </body>
 </html>
